Pushed audit-trigger specific code into the engine
