+++ splitgraph/commands/_objects/utils.py
@@ -4,103 +4,13 @@ Internal helper functions for packaging changes recorded by the audit trigger in
 
 from random import getrandbits
 
-from splitgraph.engine import get_engine
 from splitgraph.exceptions import SplitGraphException
-
-
-def get_replica_identity(schema, table):
-    """
-    Get the PK we're storing changes as. If the table has no PK, we treat the whole row as the primary key,
-    meaning that we don't support
-    """
-    engine = get_engine()
-    return engine.get_primary_keys(schema, table) or engine.get_column_names_types(schema, table)
-
-
-def _split_ri_cols(action, row_data, changed_fields, ri_cols):
-    """
-    :return: `(ri_vals, non_ri_cols, non_ri_vals)`: a tuple of 3 lists:
-        * `ri_vals`: values identifying the replica identity (RI) of a given tuple (matching column names in `ri_cols`)
-        * `non_ri_cols`: column names not in the RI that have been changed/updated
-        * `non_ri_vals`: column values not in the RI that have been changed/updated (matching colnames in `non_ri_cols`)
-    """
-    non_ri_cols = []
-    non_ri_vals = []
-    ri_vals = [None] * len(ri_cols)
-
-    if action == 'I':
-        for cc, cv in row_data.items():
-            if cc in ri_cols:
-                ri_vals[ri_cols.index(cc)] = cv
-            else:
-                non_ri_cols.append(cc)
-                non_ri_vals.append(cv)
-    elif action == 'D':
-        for cc, cv in row_data.items():
-            if cc in ri_cols:
-                ri_vals[ri_cols.index(cc)] = cv
-    elif action == 'U':
-        for cc, cv in row_data.items():
-            if cc in ri_cols:
-                ri_vals[ri_cols.index(cc)] = cv
-        for cc, cv in changed_fields.items():
-            # Hmm: these might intersect with the RI values (e.g. when the whole tuple is the replica identity and
-            # we're updating some of it)
-            non_ri_cols.append(cc)
-            non_ri_vals.append(cv)
-
-    return ri_vals, non_ri_cols, non_ri_vals
-
-
-def _recalculate_disjoint_ri_cols(ri_cols, ri_vals, non_ri_cols, non_ri_vals, row_data):
-    # If part of the PK has been updated (is in the non_ri_cols/vals), we have to instead
-    # apply the update to the PK (ri_cols/vals) and recalculate the new full new tuple
-    # (by applying the update to row_data).
-    new_nric = []
-    new_nriv = []
-    row_data = row_data.copy()
-
-    for nrc, nrv in zip(non_ri_cols, non_ri_vals):
-        try:
-            ri_vals[ri_cols.index(nrc)] = nrv
-        except ValueError:
-            row_data[nrc] = nrv
-
-    for col, val in row_data.items():
-        if col not in ri_cols:
-            new_nric.append(col)
-            new_nriv.append(val)
-
-    return ri_vals, new_nric, new_nriv
 
 
 def _merge_changes(old_change_data, new_change_data):
     old_change_data = {k: v for k, v in zip(old_change_data['c'], old_change_data['v'])}
     old_change_data.update({k: v for k, v in zip(new_change_data['c'], new_change_data['v'])})
     return {'c': list(old_change_data.keys()), 'v': list(old_change_data.values())}
-
-
-def convert_audit_change(action, row_data, changed_fields, ri_cols):
-    """
-    Converts the audit log entry into Splitgraph's internal format.
-    :returns: [(pk, kind, extra data)] (more than 1 change might be emitted from a single audit entry).
-    """
-    ri_vals, non_ri_cols, non_ri_vals = _split_ri_cols(action, row_data, changed_fields, ri_cols)
-    pk_changed = any(c in ri_cols for c in non_ri_cols)
-    if pk_changed:
-        assert action == 'U'
-        # If it's an update that changed the PK (e.g. the table has no replica identity so we treat the whole
-        # tuple as a primary key), then we turn it into a delete old tuple + insert new one.
-        result = [(tuple(ri_vals), 1, None)]
-
-        # Recalculate the new PK to be inserted + the new (full) tuple, otherwise if the whole
-        # tuple hasn't been updated, we'll lose parts of the old row (see test_diff_conflation_on_commit[test_case2]).
-        ri_vals, non_ri_cols, non_ri_vals = _recalculate_disjoint_ri_cols(ri_cols, ri_vals,
-                                                                          non_ri_cols, non_ri_vals, row_data)
-        result.append((tuple(ri_vals), 0, {'c': non_ri_cols, 'v': non_ri_vals}))
-        return result
-    return [(tuple(ri_vals), KIND[action],
-             {'c': non_ri_cols, 'v': non_ri_vals} if action in ('I', 'U') else None)]
 
 
 def conflate_changes(changeset, new_changes):
@@ -133,10 +43,6 @@ def conflate_changes(changeset, new_changes):
                 if old_change[0] == 0 or old_change[0] == 2:
                     new_data = _merge_changes(old_change[1], change_data)
                     changeset[change_pk] = (old_change[0], new_data)
-
-
-# "Truncate" kind is missing
-KIND = {'I': 0, 'D': 1, 'U': 2}
 
 
 def get_random_object_id():
